change mileage to decimal

+++ tracking_mgt/models.py
@@ -5,7 +5,7 @@ class GpsDailyReport(models.Model):
     campaign_name = models.CharField(max_length=255, db_index=True, null=True, default="phd")
     created_date = models.DateTimeField(auto_now=False, db_index=True, blank=True, null=True)
     license_no =  models.CharField(max_length=15, db_index=True, null=True)
-    mileage = models.IntegerField(default=0)
+    mileage = models.DecimalField(default=0, max_digits=9, decimal_places=3)
     viewer = models.IntegerField(default=0)
     def __str__(self):
         return "%s / %s / %s"%(self.created_date, self.mileage, self.viewer)
